fix: errors with seeders

+++ database/migrations/2024_03_17_192911_create_services_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->longText('description')->nullable();
             $table->integer('approximate_duration_min')->nullable();
-            $table->decimal('price', 4)->default(0);
+            $table->float('price', 4)->default(0);
             $table->timestamps();
         });
